Update window creation and drawing

+++ Mallet/src/ui/docking/dockingnode.h
@@ -38,5 +38,5 @@ public:
 	DockingDirection direction;
 
 	float ratio;
-	float absOffset;
+	int absOffset;
 };
